Generate metadata for dmgs as well

+++ lib/omnibus/project.rb
@@ -116,6 +116,8 @@ module Omnibus
           run_pkgmk
         elsif pkg_type == 'mac_pkg'
           run_mac_package_build
+        elsif pkg_type == 'mac_dmg'
+          # noop, since the dmg creation is handled by the packager
         else # pkg_type == "fpm"
           run_fpm(pkg_type)
         end
@@ -539,9 +541,6 @@ module Omnibus
     # If specific types cannot be determined, default to `["makeself"]`.
     #
     # @return [Array<(String)>]
-    #
-    # @todo Why does this only ever return a single-element array,
-    #   instead of just a string, or symbol?
     def package_types
       case platform_family
       when 'debian'
@@ -555,7 +554,7 @@ module Omnibus
       when 'windows'
         ['msi']
       when 'mac_os_x'
-        ['mac_pkg']
+        ['mac_pkg', 'mac_dmg']
       else
         ['makeself']
       end
@@ -616,6 +615,11 @@ module Omnibus
     def render_metadata(pkg_type)
       basename = output_package(pkg_type)
       pkg_path = "#{config.package_dir}/#{basename}"
+
+      # Don't generate metadata for packages that haven't been created.
+      # TODO: Fix this and make it betterer
+      return unless File.exist?(pkg_path)
+
       artifact = Artifact.new(pkg_path, [platform_tuple], version: build_version)
       metadata = artifact.flat_metadata
       File.open("#{pkg_path}.metadata.json", 'w+') do |f|
@@ -637,6 +641,9 @@ module Omnibus
         "#{package_name}-#{build_version}-#{iteration}.solaris"
       when 'mac_pkg'
         Packager::MacPkg.new(self).package_name
+      when 'mac_dmg'
+        pkg = Packager::MacPkg.new(self)
+        Packager::MacDmg.new(pkg).package_name
       else # fpm
         require "fpm/package/#{pkg_type}"
         pkg = FPM::Package.types[pkg_type].new
